move ip validation to ip.js
