+++ lib/ip.js
@@ -1,5 +1,6 @@
 'use strict';
 
+var net = require('net');
 var ipaddr = require('ip-address');
 
 
@@ -64,4 +65,21 @@ exports.bitAt = function(rawAddress, idx) {
 
   // Shift the offset rightwards by bitIdx bits and & it to grab the bit
   return (rawAddress[bufIdx] >>> bitIdx) & 1;
+};
+
+
+exports.validate = function(ip) {
+  var version = net.isIP(ip);
+  switch (version) {
+    case 4:
+      return net.isIPv4(ip);
+    case 6:
+      return net.isIPv6(ip);
+    case 0:
+      return false;
+    default:
+      throw new Error(
+        'net.isIP call returned unexpected value: "' + version + '"'
+      );
+  }
 };
