+++ test/index.js
@@ -8,17 +8,10 @@ describe('index', function() {
 
 
   describe('validate()', function() {
-    it('should work fine for IPv4', function() {
+    it('should work fine for both IPv4 and IPv6', function() {
       assert.equal(maxmind.validate('64.4.4.4'), true);
-      assert.equal(maxmind.validate('64.4.4.boom!'), false);
-      assert.equal(maxmind.validate(undefined), false);
-      assert.equal(maxmind.validate('kraken'), false);
-    });
-
-    it('should work fine for IPv6', function() {
       assert.equal(maxmind.validate('2001:4860:0:1001::3004:ef68'), true);
-      assert.equal(maxmind.validate('::64.17.254.216'), true);
-      assert.equal(maxmind.validate('2001:4860:0:1001::3004:boom!'), false);
+      assert.equal(maxmind.validate('whhaaaazza'), false);
     });
   });
